Business logic for HomeScreen was added

+++ app/src/main/java/com/study/watertracker/domain/model/DayWaterIntake.kt
@@ -3,7 +3,7 @@ package com.study.watertracker.domain.model
 import java.util.*
 
 data class DayWaterIntake(
-    val amount: Int = 0,
+    val amount: Float = 0,
     val metric: WaterMetrics = WaterMetrics.MILLILITRES,
     val date: Calendar = Calendar.getInstance()
 )
